requirements.txt
DateTime==4.0.1
-e git://github.com/django/django.git@c047dda057542b3b8722c6433028e3b800c38af9#egg=Django-master
Markdown==2.3.1
Pillow==2.2.1
Pygments==1.6
django-filter==0.7
djangorestframework==2.3.9
lockfile==0.9.1
mimeparse==0.1.3
python-daemon==1.6
python-dateutil==2.2
pytz==2013.8
six==1.4.1
tornado==3.1.1
torneira==0.3.3
wsgiref==0.1.2
zope.interface==4.0.5